[Architecture] Refactor FusedMoE
