+++ tensorrt_llm/_torch/modules/fused_moe/__init__.py
@@ -1,0 +1,22 @@
+from .create_moe import create_moe, get_moe_cls
+from .fused_moe_cutlass import CutlassFusedMoE
+from .fused_moe_trtllm_gen import TRTLLMGenFusedMoE
+from .fused_moe_vanilla import VanillaMoE
+from .interface import MoE, MoEWeightLoadingMode
+from .moe_load_balancer import MoeLoadBalancer
+from .routing import (BaseMoeRoutingMethod, DeepSeekV3MoeRoutingMethod,
+                      DefaultMoeRoutingMethod,
+                      Llama4RenormalizeMoeRoutingMethod,
+                      LoadBalancedMoeRoutingMethod, Qwen3MoeRoutingMethod,
+                      RenormalizeMoeRoutingMethod, RoutingMethodType,
+                      SparseMixerMoeRoutingMethod, StaticMoeRoutingMethod)
+
+__all__ = [
+    "VanillaMoE", "CutlassFusedMoE", "TRTLLMGenFusedMoE",
+    "BaseMoeRoutingMethod", "MoeLoadBalancer", "Qwen3MoeRoutingMethod",
+    "Llama4RenormalizeMoeRoutingMethod", "SparseMixerMoeRoutingMethod",
+    "LoadBalancedMoeRoutingMethod", "StaticMoeRoutingMethod",
+    "DefaultMoeRoutingMethod", "DeepSeekV3MoeRoutingMethod",
+    "RoutingMethodType", "RenormalizeMoeRoutingMethod", "MoE",
+    "MoEWeightLoadingMode", "get_moe_cls", "create_moe"
+]
