+++ tests/unittest/_torch/auto_deploy/unit/singlegpu/custom_ops/test_ad_moe_op.py
@@ -4,7 +4,7 @@ import torch.nn.functional as F
 from _torch.helpers import reference_moe_torch
 
 import tensorrt_llm._torch.auto_deploy.custom_ops  # noqa: F401
-from tensorrt_llm._torch.modules.fused_moe import FusedMoE  # noqa: F401
+from tensorrt_llm._torch.modules.fused_moe import MoE  # noqa: F401
 
 
 @pytest.mark.parametrize("dtype", [torch.float16, torch.bfloat16])
