+++ tests/unittest/_torch/modules/test_fused_moe.py
@@ -15,8 +15,8 @@ from utils.util import (skip_neither_ada_nor_hopper_unittest,
 from tensorrt_llm._torch.autotuner import AutoTuner, autotune
 from tensorrt_llm._torch.model_config import ModelConfig
 from tensorrt_llm._torch.modules.fused_moe import (BaseMoeRoutingMethod,
+                                                   CutlassFusedMoE,
                                                    DefaultMoeRoutingMethod,
-                                                   FusedMoE,
                                                    RenormalizeMoeRoutingMethod,
                                                    VanillaMoE)
 from tensorrt_llm._torch.modules.gated_mlp import GatedMLP
@@ -34,7 +34,7 @@ MPI.pickle.__init__(
 
 @pytest.mark.parametrize(
     "moe_cls, dtype, experts, RoutingMethodCls",
-    product([FusedMoE, VanillaMoE], [torch.float16, torch.bfloat16],
+    product([CutlassFusedMoE, VanillaMoE], [torch.float16, torch.bfloat16],
             [3, 8, 512],
             [DefaultMoeRoutingMethod, RenormalizeMoeRoutingMethod]))
 def test_fused_moe(moe_cls, dtype, experts, RoutingMethodCls, mapping=None):
@@ -107,7 +107,7 @@ def test_fused_moe(moe_cls, dtype, experts, RoutingMethodCls, mapping=None):
 
 @pytest.mark.skipif(torch.cuda.device_count() < 4,
                     reason="needs 4 GPUs to run this test")
-@pytest.mark.parametrize("moe_cls", [FusedMoE, VanillaMoE])
+@pytest.mark.parametrize("moe_cls", [CutlassFusedMoE, VanillaMoE])
 @pytest.mark.parametrize("ep_size", [1, 2, 4])
 def test_fused_moe_multi_gpu(moe_cls, ep_size):
     world_size = 4
@@ -176,13 +176,14 @@ def test_fused_moe_fp8(dtype):
         weights[f"{expert_id}.w3.input_scale"] = w3_input_scale
 
     quant_config = QuantConfig(quant_algo=QuantAlgo.FP8)
-    fused_moe = FusedMoE(num_experts=NUM_EXPERTS,
-                         routing_method=routing_method,
-                         hidden_size=HIDDEN_SIZE,
-                         intermediate_size=INTERMEDIATE_SIZE,
-                         dtype=dtype,
-                         reduce_results=False,
-                         model_config=ModelConfig(quant_config=quant_config))
+    fused_moe = CutlassFusedMoE(
+        num_experts=NUM_EXPERTS,
+        routing_method=routing_method,
+        hidden_size=HIDDEN_SIZE,
+        intermediate_size=INTERMEDIATE_SIZE,
+        dtype=dtype,
+        reduce_results=False,
+        model_config=ModelConfig(quant_config=quant_config))
     fused_moe.cuda()
     fused_moe.load_weights([weights])
 
@@ -279,13 +280,14 @@ def test_fused_moe_nvfp4(dtype):
         weights[f"{expert_id}.w3.weight_scale_2"] = 1.0 / w3_w1_global
 
     quant_config = QuantConfig(quant_algo=QuantAlgo.NVFP4)
-    fused_moe = FusedMoE(num_experts=NUM_EXPERTS,
-                         routing_method=routing_method,
-                         hidden_size=HIDDEN_SIZE,
-                         intermediate_size=INTERMEDIATE_SIZE,
-                         dtype=dtype,
-                         reduce_results=False,
-                         model_config=ModelConfig(quant_config=quant_config))
+    fused_moe = CutlassFusedMoE(
+        num_experts=NUM_EXPERTS,
+        routing_method=routing_method,
+        hidden_size=HIDDEN_SIZE,
+        intermediate_size=INTERMEDIATE_SIZE,
+        dtype=dtype,
+        reduce_results=False,
+        model_config=ModelConfig(quant_config=quant_config))
     fused_moe.load_weights([weights])
     fused_moe.cuda()
 
@@ -368,13 +370,14 @@ def test_fused_moe_w4afp8(dtype):
         weights[f"{expert_id}.w3.input_scale"] = w3_input
 
     quant_config = QuantConfig(quant_algo=QuantAlgo.W4A8_AWQ)
-    fused_moe = FusedMoE(num_experts=NUM_EXPERTS,
-                         routing_method=routing_method,
-                         hidden_size=HIDDEN_SIZE,
-                         intermediate_size=INTERMEDIATE_SIZE,
-                         dtype=dtype,
-                         reduce_results=False,
-                         model_config=ModelConfig(quant_config=quant_config))
+    fused_moe = CutlassFusedMoE(
+        num_experts=NUM_EXPERTS,
+        routing_method=routing_method,
+        hidden_size=HIDDEN_SIZE,
+        intermediate_size=INTERMEDIATE_SIZE,
+        dtype=dtype,
+        reduce_results=False,
+        model_config=ModelConfig(quant_config=quant_config))
     fused_moe.load_weights([weights])
     fused_moe.cuda()
